rename ruma_wrapper to router
